requirements.txt
pytest~=7.4.3
allure-pytest